First draft integrator for 1 dimension chain
A hard-coded example of a third derivative system in which the output
dimension is 1 is given. In this example, the initial state is
(5, 5, 5), as assigned at line 57, and a state-feedback controller is
applied on line 108, as found by control.lqr() from the Python Control
System Library using identity cost matrices.

dynamaestro is a valid ROS package. It has been tested on ROS Indigo
with the "catkin" build system. Create and build it as usual; try

  roslaunch dynamaestro trivial-demo.launch

The x, y, z Point variables correspond to the state vector.