Basic implementation of general case of chain of integrators
without noise or disturbances. It is provided by class TrajectoryGenerator
and may later be moved to a class with a more focused name that inherits
from TrajectoryGenerator somehow.

The class changed substantially from the previous commit. Modulo change
to class name, the methods and other aspects should remain mostly as is,
until tests begin involving other components.

The same example as in the previous two commits is still operational
using the roslaunch file. Internally the only hardcoded parts are the
initial state to (5.0, 5.0, 5.0) and the state-feedback controller, which
will next be removed.