Use default "base_prefix" for Kobuki Gazebo plugin
Before this changeset, it was assigned the empty string, which causes
an unintentional move of several topics to outside the vehicle
namespace.

Before this changeset, the described behavior did not occur when using
an extended version of the Kobuki Gazebo plugin.  There is a pull
request <https://github.com/yujinrobot/kobuki_desktop/pull/40> that
provides the feature of allowing empty base_prefix. However, it may
never be accepted upstream, so we now prefer to have some prefix.