Dump formula in SPIN LTL syntax; does not include Xinit, Y, U
The output and input bounds and the initial state may later be
included as part of the formula. Including the output bound (Y) and
the input bound (U) is trivial if they are regarded as labeled
polytopes. However, the initial state is a point, so some tolerance of
occupancy may be needed if the same labeling infrastructure will be
used for it.