Improve documentation; change random to Polytope::randomH()
Polytope::randomH(), formerly known as Polytope::random(), now samples
matrices with n+1 rows, allowing for simplices. The renaming is to
emphasize half-space representation and in anticipation of randomV(),
which will sample vertices and then compute the convex hull. It may
eventually be interesting to provide the number of rows as a parameter
to randomH().