Create dynamaestro/logger for transcription to standard formats
For demo, try `roslaunch fmrb_sci_examples lqr.launch`. The result of
transcription is the topic /logger/output_PointStamped. E.g., it can
be viewed in rviz.

The mapping of states to geometry_msgs::Point is indicated in a log
message of the form dynamaestro_logger: Using the mapping (0,1,2) -> (x,y,z)

If the dimension of the output space is less than 3, then remaining
coordinates are 0-filled. Currently the choice of mapping is
hard-coded in log.cpp, but it is easily changed by modifying the
instantiation of DMTranscriber in log.cpp.