Naive example controller for first domain; new Python package
The example controller for the first domain visits all goals in order
by repeated construction and application of LQR solutions to stabilize
about the mean point in each "goal" polytope. Currently it ignore
obstacles, but some easy method to go around them will be added.

This changeset also introduces the fmrb Python package, which will
both provide common and domain-specific routines and representations
for reviewing and working with the benchmarks, but which is not
intended to be part of the benchmarks per se.