For first domain, dm publishes and latches init state, waits
...instead of repeatedly publishing the initial state until the first
input arrives, as was the case before this changeset. The lqr example
has been updated accordingly, and in doing so, a race condition was
revealed in the lqr example controller during the creation of
connections to the "output" and "input" topics.

This new design has the disadvantage of an arbitrarily large
difference between times in the headers of the first two state
messages; thus the first time difference is not related to the
discretization period. However, it has the advantage of unambiguous
association of the first input and the second state, which results
from its application, and the beginning of dynamics simulation.