Post instances to a topic; trial end indicated by empty state
Two major changes to the first domain (scaling chains of integrators)
for monte carlo controller evaluation: generated problem instances are
advertised via a ROS topic, not via the Parameter Server; and trial
termination is indicated by an empty state vector.

The motivation for the first change is to improve performance and to
remove ambiguity about when the instance changes. The second change
provides a reliable indicator of trial termination without the need
for separate media, thus simplifying client implementations, provided
that the output dimension of this problem domain is never zero when
using monte carlo evaluation.