User-configurable random problem instance and trial parameters
...depending on values registered with the ROS Parameter Server.  Most
are simply range of the form "J K", but the specification of output
and input bounds are one of two forms:

1. "J", which causes a cube of side length 2*J and center at the
   origin to be used;

2. "x1_min x1_max x2_min x2_max ...", which has the same
   interpretation as the bounds parameter of Polytope::box().

In this changeset, several parameters of Problem::random() have
significantly changed. In summary the support for sampling Xinit can
optionally be restricted, and the polytopes Y (output) and U (input)
are fixed, not randomly chosen. Consult the API manual for details.