Trial orchestration from a script, but not yet saving results
for the first domain. There is now a JSON format referred to as
"trialconf" that defines the configuration for a run of the dm,
including specification of the pdf supports for various parameters
involved in creating random problem instances. It summarizes important
parts of lqr.launch from before this changeset.

Let $FMRBENCHMARK be the path to your copy of the fmrbenchmark repo.
As of this commit, to run the LQR controller example with the same
configuration as before this changeset, use the commands

  $FMRBENCHMARK/domains/integrator_chains/trial-runner.py `rospack find sci_concrete_examples`/trialconf/mc-small-out3-order3.json |roslaunch -

  roslaunch sci_concrete_examples lqr.launch