Delete vestigial Stage usage of e-agents/wander
We will use more sophisticated simulation engines than Stage, so there
are no plans to maintain Stage support. The traces are "vestigial"
because Stage was used in the separate, educational context from which
e-agents/wander originates.

Stage could be re-introduced later, e.g., if its simplifications are
justified to explore simulations with a very large number of robots.