e-agent wander moves along single road segment
e-agent wander now reads a road network description (RND) file from
the command-line, instead of a random number seed. In this changeset
the PositionR2 class is deleted in favor of Eigen::Vector2d, which is
much more useful. Some memory management cruft is deleted; it was not
necessary, independently of the introduction Eigen.

    $FMRBENCHMARK/domains/dubins_traffic/trial-runner.py $(rospack find dubins_traffic_utils)/examples/data/square.json
    roslaunch wander adversaries.launch