base segment coordinates are integers; better e-agent "wander"
Base segment coordinates are required to be integers. This is without
loss of generality given the scaling length and transforms, and
assuming road segment endpoints that are rational numbers.

e-agent "wander" moves along road segments. It follows the center
line.  Thus, because roads have two-lanes, it is not obeying traffic
rules. The motion is not yet randomized, and it is still possible for
there to be consecutive waypoints not corresponding to a segment.