Init ROS message type for labels in dubins_traffic
It is copied from domains/integrator_chains/integrator_chains_msgs/msg/LabelStamped.msg
as of 8b7f144caf0eb6d5e5dca5ad8279310157ae093c.

The motivations for copying instead of depending on
integrator_chains_msgs are to avoid dependencies between problem
domain implementations for items that are fairly simple, and to
prepare for anticipated changes of the message type for labeling
states as practical experience is gained with dubins_traffic.