goals in canonical coordinates in dubins_traffic instance JSON
Before this changeset, goal intersections were listed as strings,
e.g., ego_t_0_0, which corresponds to the label when the ego vehicle
is sufficiently near the intersection 0,0. Now, the goal intersections
are expressed as arrays in JSON, e.g., [0,0], because the string label
is easy to construct if given the numerical coordinates.
Furthermore, giving goals in canonical coordinates facilitates other
operations, e.g., mapping to the actual position of the intersection
using methods of the RoadNetwork class.