going forward, license is Apache 2.0
keeping in mind that BSD 3-clause remains for work from years before
2017. This model balances our need for progress with a license that
accounts for patents, while not being annoying about prior code.

There is a legal question about when new contributions are "derivative
work" and therefore subject to the BSD 3-clause. Nonetheless, if the
new code is regarded as "derivate work", then is the question of it
being subject to a new patent relevant?

https://github.com/fmrchallenge/fmrbenchmark/issues/15