CI: build googletest on Ubuntu
This changeset adds a link to pthread in test executables because it
appears necessary on Ubuntu hosts where googletest must be built.

Before and after this change (i.e., independently of this changeset),
I succeed in building and running the test on Fedora 32 (GNU/Linux),
where gtest is a shared library installed via the Fedora package manager.